refactor: consolidate endpoints config logic to dedicated getter
