+++ api/server/services/Config/index.js
@@ -3,10 +3,9 @@ const getCustomConfig = require('./getCustomConfig');
 const loadCustomConfig = require('./loadCustomConfig');
 const loadConfigModels = require('./loadConfigModels');
 const loadDefaultModels = require('./loadDefaultModels');
+const getEndpointsConfig = require('./getEndpointsConfig');
 const loadOverrideConfig = require('./loadOverrideConfig');
 const loadAsyncEndpoints = require('./loadAsyncEndpoints');
-const loadConfigEndpoints = require('./loadConfigEndpoints');
-const loadDefaultEndpointsConfig = require('./loadDefaultEConfig');
 
 module.exports = {
   config,
@@ -16,6 +15,5 @@ module.exports = {
   loadOverrideConfig,
   loadAsyncEndpoints,
   ...getCustomConfig,
-  loadConfigEndpoints,
-  loadDefaultEndpointsConfig,
+  ...getEndpointsConfig,
 };
